[windrbd] windrbd-root setting
This is the DRBD Url we pass to the WinDRBD driver so that
it connects to the root DRBD device ('C:' drive). Next step
would be to actually use that setting (instead of the hardcoded
setting). Setting the windrbd-root variable also works via
DHCP. To enable the setting, add:

option ipxe.windrbd-root code 196 = string;

and later something like:

option ipxe.windrbd-root "drbd:tiny-windows-disk;C;2;0.0.0.0:7683;1;1;johannes-VirtualBox;1;192.168.56.102:7683";

to your /etc/dhcp/dhcpd.conf

Since we boot a WinDRBD enabled Windows via http we cannot just
use the sanboot URL for the parameters.