[windrbd] Have feature flag for WinDRBD support.
This allows us now to directly test for windrbd support
being present as follows in the /etc/dhcp/dhcpd.conf:

        if exists ipxe.windrbd {
		filename "";
                option root-path "http://example.com/cgi-bin/drbd.cgi?DRBD_MINOR=123";
		option ipxe.windrbd-root <WinDRBD-boot-URL>;
	} else {
		filename "http://192.168.56.102/ipxe/ipxe-windrbd.pxe";
	}

This was formerly done by checking for iSCSI support (which
happened to work on my machine). From now on, use ipxe.windrbd
to test for WinDRBD support.

The feature flag and the WinDRBD root URL must be defined as
follows in the /etc/dhcp/dhcpd.conf:

  option ipxe.windrbd code 42 = unsigned integer 8;
  option ipxe.windrbd-root code 196 = string;